Review data integrated & first DecisionTree created
INTEGRATED REVIEW DATA:

Yesterday, the programs that amend various features in the critic_ratings table (including one that adds on the Ebert ratings) were functionally complete, but disorganized. Today, those were assembled together as the class remedyReviewTable, which is in the review_remedies.py file.

This class connects to the database after it's given its path upon instantiation. It then has separate methods for filling in missing Metacritic values; filling in missing RT values; and adding on the Ebert ratings. The class also has a method for disconnecting from the MySQL db and shutting down the SQLAlchemy engine.

The script 'amend_CR_table.py' leverages this class to execute such amending of the critic_ratings table.

ANALYSIS BEGINS - FIRST DECISION TREE

With all that done, data only needed slightly more prep before I could train a first Decision Tree on it. Though the accuracy of 0.75 was moderately satisfying, the modeling has so far been disappointing. I put extensive work into the completeness and richness of this critic_ratings table, but the modelling so far seems to have no need for it: the simple question of whether a film has a Metacritic review of 0.89 is what determines if it gets any 'Awesome-ness' in its rating. The three other review score attributes add only an exceedingly negligible benefit.

It looks like genre might be key to refining the Decision Tree further. This is my impression after glancing at the samples that fill leaf node #4, an impure one with many samples.