Created "reconcile_dup_entries()" method
Created this method to deal with entries that share a title and director (which I interpret as reflecting the same film.) The older entry is considered the point of fidelity.

It could be worthy to write this method's reporting to a log file. It reports the differences between all the pairs of records it considers.