Overhaul mc_search_and_scrape
I have added functionality to mc_search_and_scrape so that it now consults master files. Whenever the method is called and it scrapes an unencountered movie, it records the scrape's output in the master files, in addition to its usual output.

If a movie's title, year, and director already show in any of the master files, the process that corresponds to that master file is skipped.