Fix the Siskel scheduler
This was necessary because the data object in which the showtimes were stored changed in structure, from a dictionary of dictionaries to a list of dictionaries (I think.) Either way, since dataframes thereof are also created and saved, I've changed the scheduler to instead work with the various dataframes as inputs.

This was a sizable overhaul of that scheduling process.