Update musicbox_scrape
Updated musicbox_scrape and scheduler with the following changes:
1. The scheduler now operates off the df of showtimes, instead of a dictionary.
2. The scheduler now incorporates the musicbox_show_info in the same way that the Siskel one does with its corresponding show info.
3. If ran as a test, the musicbox_scrape now creates and outputs to a subdir named 'test'.

The scrapers and schedulers for the Musicbox and Siskel are now recognizably similar enough logically that I should maybe strive to create a single scraper method and single scheduler that works for both. I imagine this would involve considerable refactoring, mostly to refactor the scraping and crawling into functions.

Doing so would make it way easier to improve both functions, because I wouldn't have to double-enter any process improvements, by applying such improvements to the methods of each theater separately.