Fixed up schedulers & started web page
Made big changes to the schedulers so that they no longer delete past showtimes. They now operate by deleting from the calendar all future showtimes; forming a dataframe of its showtimes still slated (which must be in the past); removing any of those showtimes from the master dataframe of scraped showtimes; slating all those showtimes as events.

It's occuring to me that my scrapers now add to a master list at each scrape, and when they do so, they are not scanning the master list to remove future showtimes from it (before adding itself on.) This is an issue I should resolve, so showtimes that get taken down by theaters don't linger in my schedules.

Aside from all this: I started up a web page to view my movie showtime data. It's daunting, but it's been helpful in clarifying what my data might need to look like in my database, to best suit the site and my vision for it.