Upd musicbox scrape - UTC-string showtimes
Now scraping and saving the showtimes as UTC-strings that clarify CDT as the timezone, instead of simply saving timezone-naive datetime objects as I was before.

The database loader(s) will need to adjust, to convert these UTC-strings to datetimes for loading perhaps. Datetimes work well for MySQL databases. However, UTC-strings are generally better for storage and readability, so I made that change.

Also, the Siskel showtimes are in UTC-strings on their page where I scrape them, so this might simply things across scrapers.