Debugged siskel double-month dupes
If a Siskel screening was shown in the "Next month" page within the few previous month's days which are sometimes featured there, then those screenings were getting double-scraped and double-scheduled.

I have made a change in the Siskel scraper so that before adding a showtime dictionary to the list of them, it checks to see if that dictionary already exists there.