Side-by-side calendar and table
I made the calendar and table side-by-side now. And they dynamically resize.

I need to mess with the header, stick it above the table and let the calendar take up the full vertical space maybe.

I added a day view, which was trivial to do. It seems convenient enough that I might want to have the table/report switch to summarizing just that day's movies should the day view be clicked.